+++ packages/default-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # fuzd-default-client
 
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies
+  - fuzd-scheduler@0.2.3
+  - fuzd-executor@0.1.10
+
 ## 0.1.15
 
 ### Patch Changes