+++ packages/tlock-decrypter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # fuzd-tlock-decrypter
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - fuzd-scheduler@0.2.3
+
 ## 0.2.2
 
 ### Patch Changes